[remove] render.js 내부의 함수들을 컴포넌트로 분리 - render.js에는 ‘웹툰-홈’ 화면과 ‘웹툰-요일연재’ 화면을 랜더링하는 함수만 남겨놓았다.
